# [SPARK-32703][SQL] Replace deprecated API calls from SpecificParquetRecordReaderBase