@sunchao 2020-08-29
> Parquet reader is performance-wise important component in Spark SQL. We better to make sure no performance regression due to this change. Should we run a benchmark to check it?

Sure let me do that. Should I just run `DataSourceReadBenchmark`?