@sunchao 2020-08-31
Thanks @HyukjinKwon - this is good to know. Since it seems Spark could take a while to get to 1.11, should I close this for now then?