@HyukjinKwon 2020-08-31
@sunchao, I would just close for the time being, and bring it back when the upgrade is done. But I don't mind leaving it open if you prefer it for any reasons.