@sunchao 2021-01-29
Re-open this since Spark has upgraded to Parquet 1.11 :-)