@srowen 2021-01-31
This can go ahead because of the Parquet 1.11.1 update?