@HyukjinKwon 2021-01-31
Yeah, because there was a resource leak problem (https://github.com/apache/spark/pull/29542#pullrequestreview-478269264) that's fixed in Parquet 1.11.