@HyukjinKwon 2021-02-24
Reverted at https://github.com/apache/spark/commit/80bad086c806fd507b1fb197b171f87333f2fb08