@srowen 2021-02-24
That's fine, but the benchmarks here seemed to show it was an improvement at least for the tests that were run. I wonder what TPC-DS covers that is different.