@HyukjinKwon 2021-02-24
Yeah, it's fine. Thanks for taking a look @sunchao. We should get rid of this deprecated usage eventually in any event.