@sunchao 2021-02-26
I tried @LuciferYang 's suggestion with TPC-DS benchmark and it fixed the perf regression. For instance, in the q9 above, here's what I got:

without the PR | with the PR | with the PR + requestSchema fix
|-----|-------|------------------------------------------------------
5940 | 11260 | 5281

I've gathered all the results in this [gist](https://gist.github.com/sunchao/533c1c9a5fb6adc5f7c72b1c465e974d). The benchmark was run with scale factor 5.

@maropu @HyukjinKwon @srowen let me know what you think. I can update this with the fix.