@maropu 2021-02-26
Thanks, @sunchao! After the fix applied, I'll check the performance on my env, again.