@sunchao 2021-02-26
Opened #31667 with the fix. Comparing to the old PR there's only one line change:
```scala
reader.setRequestedSchema(requestedSchema);
```
Please take a look. Thanks!